fix: 금액 검증 로직 이동
